+++ src/events.js
@@ -1,4 +1,7 @@
 export default {
-  REQUEST_INITIAL_THEME: 'storybook-theming-addon/request-initial-theme',
   INITIAL_THEME: 'storybook-theming-addon/initial-theme',
+  DECORATOR_MOUNTED: 'storybook-theming-addon/decorator-mounted',
+  DECORATOR_UNMOUNTED: 'storybook-theming-addon/decorator-unmounted',
+  REQUEST_INITIAL_THEME: 'storybook-theming-addon/request-initial-theme',
+  UPDATE_THEME: 'storybook-theming-addon/update_theme',
 };